% Auto-generated: do not edit by hand
\name{button}

\alias{button}

\title{button component}

\description{

}

\usage{
button(children=NULL, id=NULL, appearance=NULL, className=NULL,
disabled=NULL, iconAfter=NULL, iconBefore=NULL, intent=NULL,
isActive=NULL, isLoading=NULL, label=NULL, n_clicks=NULL,
size=NULL, value=NULL)
}

\arguments{
\item{children}{Character. }

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{appearance}{Character. }

\item{className}{Character. }

\item{disabled}{Logical. }

\item{iconAfter}{Dash component. }

\item{iconBefore}{Dash component. }

\item{intent}{Character. }

\item{isActive}{Logical. }

\item{isLoading}{Logical. }

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{n_clicks}{Numeric. }

\item{size}{A value equal to: "small", "medium", "large". Dash-assigned callback that should be called to report property changes
to Dash, to make them available for callbacks.}

\item{value}{Character. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

